consolidating to use LogEntry
